implemented messageDeleting service added it to session service when deleting also added deleting method to bot and little type refactor

+++ wrapper-ripple/src/main/java/com/hatiko/ripple/telegram/bot/database/service/XrpDatabaseOperator.java
@@ -24,8 +24,8 @@ public interface XrpDatabaseOperator {
 	
 	Boolean deleteAccountLastIdByPublicKey(String publicKey);
 	
-	MessageIdDTO getMessageId(Integer chatId);
-	MessageIdDTO updateMessageId(Integer chatId, Integer lastSent, Integer lastDeleted);
+	MessageIdDTO getMessageId(Long chatId);
+	MessageIdDTO updateMessageId(Long chatId, Integer lastSent, Integer lastDeleted);
 	Boolean deleteMessageId(Integer chatId);
 
 }
