+++ wrapper-ripple/src/main/java/com/hatiko/ripple/telegram/bot/handler/operations/GetTransactionInfoMessageHandler.java
@@ -48,6 +48,6 @@ public class GetTransactionInfoMessageHandler implements TelegramMessageHandler 
 		}
 		
 		Integer sentMessageId = responseMessageOperator.responseGetTransactionInfo(null, chatId, 0);
-		databaseOperator.updateMessageId((int)(long)chatId, sentMessageId, null);
+		databaseOperator.updateMessageId(chatId, sentMessageId, null);
 	}
 }
